Add Intro to Deep Learning with PyTorch (#50)
Add "Intro to Deep Learning with PyTorch", a free course by Udacity and Facebook AI, which uses PyTorch to implement deep neural networks. It also has an interview by Soumith Chintala, one of the creators of PyTorch. It covers an Intro to PyTorch, and also a tutorial for deploying it to C++.