Added albumentations library (#64)
I am proposing to add Albumentations library to the list. https://github.com/albu/albumentations

The library is used in many companies for Computer Vision tasks
And in a Computer Vision competitions at kaggle.com and other platforms.

It is actively maintained and new features are added on a regular basis. The repo currently has 2000+ stars.

I think it would be a great addition to the `awesome-pytorch-list`.